Removed deprecated code
IntelliJ was letting me know the use of new(Long (n)) was deprecated, so I updated it in order to address technical debt. This relates to a section within the article Clean Coding in Java where it says "Maintainable Codebase: Any software that we develop has a productive life and during this period will require changes and general maintenance. Clean code can help develop software that is easy to change and maintain over time" (Chandrakant). This relates to the project as removing deprecated features helps the code stay maintained.