Resolved build errors
This commit was mostly to resolve build errors I was getting when trying to build your project. In Java, List is abstract so you cannot instantiate it in the way attempted. I also added the @Nullable annotation to FacilityServiceStub as it is recommended by the article "A short summary of Java coding best practices". It says "Encourage use of @Nullable when a field or method returns null" which it did (Hamedy).