Corrected href
In this commit I simply corrected your href as it was pointing to the wrong html file and I also made the link in the page say "Facility" instead of faculty as it seems this is what was desired. Removing unused code, like the namespace at the top, is something that also keeps your code clean and readable per the articles addressing clean code.