Fixed grammatical error on webpage.
Per Hamedy in the "A short summary of Java coding best practices" article, coding standards are very important in making sure all users and developers can read or change code. While this is a very small mistake, I believe a user would pick up on it very fast in the user interface. This ties into making sure your code is clean and readable for all users.